Move 3D code to ScottPlot.ScottPlot3D
