+++ src/ScottPlot5/Sandbox.Avalonia3D/Views/MainWindow.axaml.cs
@@ -1,5 +1,6 @@
 using Avalonia.Controls;
-using Sandbox.Avalonia3D.Primitives3D;
+using ScottPlot.ScottPlot3D.Plottables3D;
+using ScottPlot.ScottPlot3D.Primitives3D;
 
 namespace Sandbox.Avalonia3D.Views;
 
@@ -9,7 +10,7 @@ public partial class MainWindow : Window
     {
         InitializeComponent();
             
-        Plottables3D.Scatter3D scatter = new();
+        Scatter3D scatter = new();
         AvaPlot.Plot3D.Plottables.Add(scatter);
 
         Point3D location = new(.25, .25, 0);
