+++ src/ScottPlot5/ScottPlot5/ScottPlot3D/IPlottable3D.cs
@@ -1,4 +1,4 @@
-namespace Sandbox.Avalonia3D;
+namespace ScottPlot.ScottPlot3D;
 
 public interface IPlottable3D
 {
