+++ src/ScottPlot5/ScottPlot5/ScottPlot3D/Plottables3D/Line3D.cs
@@ -1,7 +1,6 @@
-using Sandbox.Avalonia3D.Primitives3D;
-using ScottPlot;
+using ScottPlot.ScottPlot3D.Primitives3D;
 
-namespace Sandbox.Avalonia3D.Plottables3D;
+namespace ScottPlot.ScottPlot3D.Plottables3D;
 
 public struct Line3D : IPlottable3D
 {
