+++ src/ScottPlot5/ScottPlot5/ScottPlot3D/Plottables3D/Scatter3D.cs
@@ -1,10 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Sandbox.Avalonia3D.Primitives3D;
-using ScottPlot;
+using ScottPlot.ScottPlot3D.Primitives3D;
 
-namespace Sandbox.Avalonia3D.Plottables3D;
+namespace ScottPlot.ScottPlot3D.Plottables3D;
 
 public class Scatter3D : IPlottable3D
 {
