+++ src/ScottPlot5/ScottPlot5/ScottPlot3D/Primitives3D/Line3D.cs
@@ -1,3 +1,3 @@
-namespace Sandbox.Avalonia3D.Primitives3D;
+namespace ScottPlot.ScottPlot3D.Primitives3D;
 
 public record struct Line3D(Point3D Start, Point3D End);
