+++ src/ScottPlot5/ScottPlot5/ScottPlot3D/Primitives3D/Point3D.cs
@@ -1,6 +1,4 @@
-using System;
-
-namespace Sandbox.Avalonia3D.Primitives3D;
+namespace ScottPlot.ScottPlot3D.Primitives3D;
 
 public record struct Point3D(double x, double y, double z)
 {
