+++ src/ScottPlot5/ScottPlot5/ScottPlot3D/Primitives3D/Rotation3D.cs
@@ -1,4 +1,4 @@
-namespace Sandbox.Avalonia3D.Primitives3D;
+namespace ScottPlot.ScottPlot3D.Primitives3D;
 
 public record struct Rotation3D
 {
