+++ src/ScottPlot5/ScottPlot5/ScottPlot3D/Primitives3D/Size3D.cs
@@ -1,3 +1,3 @@
-namespace Sandbox.Avalonia3D.Primitives3D;
+namespace ScottPlot.ScottPlot3D.Primitives3D;
 
 public record struct Size3D(double Width, double Length, double Height);
